Item Reader preparation workspace

+++ src/main/java/com/shindhe/writer/FirstItemWriter.java
@@ -6,9 +6,9 @@ import org.springframework.stereotype.Component;
 import java.util.List;
 
 @Component
-public class FirstItemWriter implements ItemWriter<Long> {
+public class FirstItemWriter implements ItemWriter<Integer> {
     @Override
-    public void write(List<? extends Long> list) throws Exception {
+    public void write(List<? extends Integer> list) throws Exception {
         System.out.println("Inside Item Writer");
         list.stream().forEach(System.out::println);
     }
